+++ client/src/api/user.js
@@ -1,7 +1,7 @@
 import { post, remove } from './base'
 
-export function login (email, password) {
-  return post('/user/login', { email, password })
+export function login (email, password, rememberMe) {
+  return post('/user/login', { email, password, remember_me: rememberMe })
 }
 
 export function deleteUser (id) {